content/canvas/test/test_2d.gradient.object.return.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.gradient.object.return</title>
<script src="/MochiKit/MochiKit.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>
var _deferred = false;

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

var g1 = ctx.createLinearGradient(0, 0, 100, 0);
ok(Object.prototype.toString.call(g1) === '[object CanvasGradient]', "Object.prototype.toString.call(g1) === '[object CanvasGradient]'");
ok(g1.addColorStop !== undefined, "g1.addColorStop !== undefined");

var g2 = ctx.createRadialGradient(0, 0, 10, 0, 0, 20);
ok(Object.prototype.toString.call(g2) === '[object CanvasGradient]', "Object.prototype.toString.call(g2) === '[object CanvasGradient]'");
ok(g2.addColorStop !== undefined, "g2.addColorStop !== undefined");

if (!_deferred) SimpleTest.finish();
});
</script>
